[#22601] xCluster: Fail setup if stream update fails
Summary:
Fail setup if we fail to update the xcluster stream state to `ACTIVE` at the end of xCluster setup.
On Bootstrap the streams are created in `INITIATED` state, and while in this state tablet splitting is disabled on the table. Also CdcService on source cluster does not cache streams that are not in `ACTIVE` state causing it to repeatedly call `GetCDCStream` on Source yb-master.

Fixes #22601
Jira: DB-11511

Test Plan: XClusterTest.FailedSetupStreamUpdate

Reviewers: jhe, xCluster

Reviewed By: jhe

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D35475