[PLAT-13987] Add ability to skip bootstrapping
Summary:
**Context**
We would like to give some advanced users with special use cases the ability to skip the bootstrapping (backup & restore)
step when they have a legitimate reason to do so.

c1fd0429ccc6e4fb59a18dd3902584fdaf3a4451 added a new boolean field in the Bootstrap params called `allowBootstrap` which
simplifies the request body for clients. Instead of requiring the user to specify exactly which tables they wish to allow
bootstrapping for, the YBA backend's new `allowBootstrap` field allows us to simply allow bootstrapping for all tables and
databases in the universe. On the UI and docs we already inform the user that bootstrapping (aka creating a full copy) is
done at the database level for YSQL. Since the user is already aware and continued to submit the form anyways, we can just
convey that intention simply by passing `allowBootstrap` = true to the backend.

**Changes**
- Add a new runtime config flag called `yb.ui.xcluster.enable_skip_bootstrapping`. When set to true,
  the user will see a secondary button appear at the bottom of the create xCluster modal as well as
  the bottom of the select tables modal (when used for xCluster replication and not xCluster DR).
  This 'Skip Creating Full Copy' button will allow the user to submit their request without specifying
  backup storage config parameters. This will cause the YBA backend to skip the backup & restore step.

- Instead of just passing tables in the `bootstrapParams`, the YBA UI will now always pass `allowBootstrapping` = true.
  We already inform the user that we will bootstrap the entire database for YSQL and the user is okay with this.
  Specifying the whole tables list again under `bootstrapParams` doesn't need to be done anymore now that the backend
  lets us specify that the user is okay with bootstrapping whatever is needed. We do continue passing a non-empty tables list
  because `tables` is still a required field.

- Minor styling changes applied to the shared YBModal component. The `footerAccessory` prop will now expand
  to fit the remaining space in the footer.

Test Plan:
- Create two universes. On the desired source universe, add the following:
   - At least one database with tables containing no data
   - At least one database with tables containing data
   Create the same databases and tables on the target universe.
- Create xCluster config

Verfiy that the user is able to skip bootstrapping when `yb.ui.xcluster.enable_skip_bootstrapping` is true.
{F255505}
{F255506}
Verify that the 'skip creating full copy' button is not shown for DR.
{F255507}
Verify that the 'skip creating full copy' button is not shown when `yb.ui.xcluster.enable_skip_bootstrapping` is false.
{F255508}
- Check the form handles back and forwards navigation after the skip bootstrapping button is pressed.
  (ex. verify that the back button on the next page doesn't bring the user to the configure bootstrap page)

Reviewers: cwang, hzare, vbansal, rmadhavan

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35768