[#21600, #21619, #21621] [DB Clone][YSQL]: Add support for DB cloning to a time before table drop
Summary:
Cloning a database to a time before drop table was failing in YSQL due to an error in generating the schema as of time t.
More specifically, `ysql_dump` fails in case the `read-time` option is used to dump the schema as of time `t`and `t` is a time before a drop table happens. This is due to bad interaction between `--include-yb-metadata` and `--read-time` options. In fact, part of including YB metadata means executing a select query `select * from yb_table_properties('table_name'::regclass);`. This in turn sends a `GetTableSchema` rpc and subsequently `GetTableLocations` RPCs to the master leader.
The master leader returns: `The object 'table_name' is not running` error as the table is in a hidden state (dropped tables covered by a snapshot schedule are marked as hidden).
Using `--include-yb-metadata` is required for backup/restore and thus for cloning, specially for the colocated case as it contains the colocation_id. Therefore, we need to make the two options work together.

Fixed by returning the `TableSchema` and `GetTableLocationsResponsePB` of the dropped table T as a response to the `GetTableSchema` and `GetTableLocations` RPCs respectively, even-though T might be marked as hidden.

The successful return of `GetTableSchema` and `GetTableLocationsResponsePB` in case of a hidden table only happens when the two fields `GetTableSchemaRequestPB.include_inactive` and `GetTableLocationsRequestPB.include_inactive` are set to true respectevly. These fields are set to true when `yb_read_time` GUC is used in a pg session and are set to false otherwise.

The field `GetTableSchemaRequestPB.include_inactive` is added in this diff in contrast to `GetTableLocationsRequestPB.include_inactive` which already existed.

The revision also fixes: #21619, #21621

**Upgrade/Rollback safety:**

Two proto messages are modified in this diff:
1- `PgOpenTableRequestPB`: a new bool field `include_inactive` is added. If the field is not set in the message (which is equivalent to setting it to false), nothing will change from the perspective of the consumer of the API. The old default behavior where include_inactive is set to false will happen in this case. I didn't add an autoflag for this field as the caller and the service reside in the same process (tserver).
2- `GetTableSchemaRequestPB`: a new bool field `include_inactive` is added.  If the field is not set in the message (which is equivalent to setting it to false), nothing will change from the perspective of the consumer of the API.
If the client (tserver) wants to set `include_inactive` to true we have two cases:
- Master leader is using the new code version: in this case the new behavior will happen as expected.
- Master leader is running an old code where `GetTableSchemaRequestPB.include_inactive` is not used:  the master returns  `OBJECT_NOT_FOUND` error. We didn't add an autoflag which could have avoided sending the `GetTableSchema` if include_inactive is set to true and the autoflag is not set to true, because we will return a  `NotSupported` error. So the autoflag is replacing one error message with another and thus deemed an overkill.

Test Plan: ./yb-build --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter PgCloneTest.CloneAfterDropTable

Reviewers: asrivastava, zdrudi, hsunder

Reviewed By: asrivastava, zdrudi

Subscribers: yguan, hsunder, yugaware, ybase, yql, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D35178