[#21534] docdb: Set owner correctly for cloned databases
Summary:
Prior to this diff, the owner of the cloned database was the same as the owner of the source database. This diff adds a modification to the `ysql_dump` output to rewrite the `ALTER ____ foo OWNER TO yugabyte` to point to the correct user (the user running the `CREATE DATABASE WITH TEMPLATE` command by default or the `OWNER` (if set).

Fixes #21534

**Upgrade/Rollback safety:**
Altered protobuf is not used in production at the moment.
Jira: DB-10420

Test Plan: `./yb_build.sh release --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter PgCloneTest.UserIsSet`

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: kramanathan, yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D35799