[#22521] DocDB: Show HIDDEN state in Master/Tserver Tables UI more prominently
Summary:
Currently the hidden state is shown in the UI as a column at the far right of the table. This is not very visible and can be easily missed. This diff moves the hidden info to the  state column to make it more visible.
Also using the new table UI helper class to make the code more consistent.

- Also moved the macros to a Class to make it simpler to use. The macro required the developer to include `HTML_END_TABLE` and `HTML_ADD_SORT_AND_FILTER_TABLE_SCRIPT`, which the classes take care of automatically, make it easier to use. And moved this class to the server folder where webserver files exist instead of the common folder.

Fixes #22521
Jira: DB-11448

Test Plan:
Manually validated the changes in the UI

Previous page:

{image uri=https://github.com/yugabyte/yugabyte-db/assets/12418230/56703f0a-e15d-4c99-81bc-5ba05a838207, width=1000px, alt=picture of a moose, href=google.com}

New page:
{image uri=https://github.com/yugabyte/yugabyte-db/assets/12418230/3d8f8604-29f2-48f5-9808-6a8fabee1f40, width=1200px, alt=picture of a moose, href=google.com}

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: ybase, esheng

Differential Revision: https://phorge.dev.yugabyte.com/D35465