[BACKPORT 2.20][#22601] xCluster: Fail setup if stream update fails
Summary:
Original commit: 1beb3a1870974a96c4ba0d823c0370573f3bbe91 / D35475
Fail setup if we fail to update the xcluster stream state to `ACTIVE` at the end of xCluster setup.
On Bootstrap the streams are created in `INITIATED` state, and while in this state tablet splitting is disabled on the table. Also CdcService on source cluster does not cache streams that are not in `ACTIVE` state causing it to repeatedly call `GetCDCStream` on Source yb-master.

Fixes #22601
Jira: DB-11511

Test Plan: XClusterTest.FailedSetupStreamUpdate

Reviewers: jhe, xCluster

Reviewed By: jhe

Subscribers: slingam, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35827