[PLAT-14316] Add generic ignore error capability in Task Executor
Summary:
This change allows ignoring a subtask failure that is specific to a task. An example is that for some task, a subtask failure is ignored but a marker is set for further processing. This change is needed for some
subsequent changes for auto master failover and other usages.

Test Plan: Added UTs.

Reviewers: cwang, sanketh, amalyshev

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35844