[#22808] CDCSDK: Prevent indexes and matview from getting added to CDC stream
Summary:
While creating a CDC stream on a database, we only add user created tables to the CDC stream metadata and filter out any existing indexes/materialized views/system tables/colocated parnt table from being added to the CDC stream metadata. But the same is not followed for any new indexes or mat views created after the stream creation and they were inadvertently getting added to the CDC stream metadata. This diff fixes that behavior and prevents any newly created indexes, mat views as well as non-user tables from getting added to the CDC stream metadata.
Jira: DB-11706

Test Plan:
Manually tested out by creating an index on an existing table as well as a dynamic table after a non-consistent snapshot CDC stream was created. Verified that it was not present in CDC stream metadata as well as CDC state table. Also, verified that it does not get added even if there is a restart or leader re-election.
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestNonUserTableShouldNotGetAddedToNonConsistentSnapshotCDCStream
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestNonUserTableShouldNotGetAddedToConsistentSnapshotCDCStream

Reviewers: xCluster, hsunder, skumar, stiwary

Reviewed By: skumar, stiwary

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35746