[#21750] YSQL: Stabilize TestPgRegressColocatedTablesWithTablespaces and TestPgRegressThirdPartyExtensionsPgCron with connection manager enabled
Summary: Within the test TestPgRegressColocatedTablesWithTablespaces and TestPgRegressThirdPartyExtensionsPgCron we override the values of the preview flags within the test, which causes the initially set value to enable connection manager to be wiped away. This patch conditionally appends the enable_ysql_conn_mgr flag to the preview flags if connection manager is being used as the endpoint.

Test Plan:
Jenkins: enable connection manager, test regex: .*testPgRegressColocatedTablesWithTablespaces.*|.*TestPgRegressThirdPartyExtensionsPgCron.*

./yb_build.sh --enable-ysql-conn-mgr-test --java-test org.yb.pgsql. TestPgRegressColocatedTablesWithTablespaces#testPgRegressColocatedTablesWithTablespaces

./yb_build.sh --enable-ysql-conn-mgr-test --java-test org.yb.pgsql. TestPgRegressThirdPartyExtensionsPgCron#schedule

Reviewers: rbarigidad, skumar, asrinivasan

Reviewed By: asrinivasan

Differential Revision: https://phorge.dev.yugabyte.com/D35846