[PLAT-14331] Add more air gap checks to Ansible install steps
Summary:
When running with an air gapped provider there are certain install steps we still execute that may
reach out to public repos to download software, which may fail if we are air gapped.

Test Plan:
Create provider with airgap enabled
Create universe, ensure the newly tagged tasks are skipped.

Reviewers: svarshney, nsingh

Reviewed By: svarshney

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D35849