[BACKPORT 2024.1][#22838] YSQL: Fix Bitmap Scan wrong results with local filters, no targets, and exceeding_work_mem
Summary:
Original commit: 2eaf0cbe9f752c591bf58c41a42c2e33bca63b68 / D35823
Issue #22622 identifies a problem where locally rechecked columns did not get targeted by Bitmap scans, so the recheck is comparing a
garbage value and the check rejects the tuple.

It was fixed by 2d12f40068d330746c85b03d0b789bcefdd591d6 / D35591, but that fix had a bug that was not detected. The bitmapset of required
columns was passed as a `Bitmapset *` instead of a `Bitmapset **`. This meant that operations that modify the bitmapset will work, but
operations that create a new bitmap set will assign their new set to the parameter but not change the original object.

This worked for the case in #22622, because the bitmap set already had a value in it (the column that was explicitly targeted). When the
bitmapset already exists, adding a new column to it modifies it.

When the bitmapset is empty, it is NULL. Adding the first value creates a new bitmapset, which was lost as soon as the bitmap went out of
scope.

```lang=c
/*
 * bms_add_member - add a specified member to set
 *
 * Input set is modified or recycled!
 */
Bitmapset *
bms_add_member(Bitmapset *a, int x)
{
    int         wordnum,
                bitnum;

    if (x < 0)
        elog(ERROR, "negative bitmapset member not allowed");
    if (a == NULL)
        return bms_make_singleton(x);

    ...
    // otherwise, mutate and return a
```
Jira: DB-11736

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressYbBitmapScans'

Reviewers: amartsinchyk, tnayak

Reviewed By: tnayak

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35888