[#22062] YSQL: Fix Bitmap Scan variable not found in subplan target list
Summary:
When a bitmap scan plan is created, the planner considers a subset of the conditions (everything
under the OR clause) to determine what could be an index condition. With YB Bitmap Scans, since we
can pushdown remote filters as well, we add a second check: if the sub-clause cannot be an index
condition, maybe it could be a remote filter.

The remote filter check needs to check that:
* the operator can be pushed down
* the clause is referring to this particular index
* all the attribute positions referenced by the condition are present in the index

Before this diff, we validated only 1 and 3. 3 only checks the attribute numbers, which may be
reused across indexes, so its still necessary to validate that the correct index is being
considered.

To help prevent future errors of this same type, rename is_index_only_refs to is_index_only_attribute_num.
This communicates its actual functionality a little bit better.
Jira: DB-10984

Test Plan:
```
./yb_build.sh  --java-test 'org.yb.pgsql.TestPgRegressYbBitmapScans'
```

Reviewers: tnayak, amartsinchyk, mtakahara

Reviewed By: mtakahara

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D35732