[#14610] DocDB: Return flag validation error message to SetFlag output
Summary:
The `SetFlag` uses `google::GetCommandLineFlagInfo` which runs the flag validation checks. If the flag validation fails, the error message is printed to the logs. This diff adds a special sink `FlagValidatorSink` and macro `LOG_FLAG_VALIDATION_ERROR` for the flag validation checks to capture the error message and return it to the caller.

Fixes #14610
Jira: DB-3981

Test Plan: FlagsTest.TestSetFlag

Reviewers: asrivastava, xCluster

Reviewed By: asrivastava

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D35841