[#22910] yugabyted: Fix metrics snapshotter is sending incorrect cpu usage metrics to system.metrics tables
Summary:
The cpu metrics sent to the `system.metrics` ycql table are incorrect, because `prev_ticks_` is not updated at every iteration.
This causes cpu metrics to be calculated using an interval between the tserver start time and the present,
rather than the previous iteration of the metrics snapshotter and the present.

The fix is simply to update `prev_ticks_` each iteration.
Jira: DB-11818

Test Plan: start up a yugabyte-db instance and compare cpu metrics in system.metrics with cpu metrics from `top`

Reviewers: nikhil, yyan

Reviewed By: nikhil, yyan

Subscribers: ybase, yugabyted-dev, djiang

Differential Revision: https://phorge.dev.yugabyte.com/D35958