[#22058] docdb: Disable connections on cloned db until cloning is complete
Summary:
This diff disallows connections on the cloned db until the clone is complete. This is done
by adding `UPDATE pg_database SET datallowconn = false WHERE datname = '<target_database>'` to
ysql_dump after connecting to the target database. Note that this is not airtight, a connection
could slip in before we connect to the target db and disallow further connections (we could enhance
this later by killing other connections if we really wanted to).

Connections are enabled once the restore is complete by running the same command with `datallowconn = true`.

**Upgrade/Rollback safety:**
The changes to an entity that is not used in production yet (SysCloneStatePB) and a brand new message (EnableDbConns).

Jira: DB-10980

Test Plan: `./yb_build.sh --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter PgCloneTest.PreventConnectionsUntilCloneSuccessful`

Reviewers: mhaddad

Reviewed By: mhaddad

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D35959