[DB-11813] Rename ysql_conn_mgr_idle_or_pending_clients metric name
Summary:
We were using the stat name **idle_or_pending_clients** which is not accurate. This stat name is misleading and hence we are moving to a different name for this stat. The new name is **waiting_clients**.

The PR makes the name change and also the changes needed for where the stat is being used.

The definition of the stat is as below

```
Number of logical connections which are either idle (i.e. no ongoing transaction) or waiting for the worker thread to be processed (i.e. waiting for od_router_attach to be called).
```
The above definition has not changed and is same as the old one.

Test Plan:
  - Hit the 13000/connections endpoint and verified that the stats are shown in accordance with the new name as **waiting_logical_connections**

  - Hit the 13000/prometheus-metrics endpoint and verified that the metrics are published in accordance the new name as **ysql_conn_mgr_waiting_clients**

Jenkins: all tests, enable connection manager

Reviewers: mkumar, rbarigidad, skumar, asrinivasan, nkumar, djiang

Reviewed By: rbarigidad, djiang

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D35992