[PLAT-14425] Fix config level replication lag query param
Summary:
The `current lag` stat on xCluster config cards and xCluster config
detail page may include lag from other tables & streams on the same
universe. This is because the query param passed was `replicationUuid`
instead of `xClusterConfigUuid`.
The metric graphs and table level replication lag stats were not
affected.

This diff fixes the query param field name.

Test Plan:
Create an xCluster config from universe A to universe B and universe A to
universe C.
Perform some writes to the tables being replicated to C and pause
universe C.
Verify that only the current lag for universe A to C increases.
{F258698}

Reviewers: rmadhavan, kkannan, vbansal, hzare

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36010