[#22893] DocDB: Use server yb client in CDCService
Summary:
Reuse the Tservers `yb::client` in `CDCService`. This will reduce 4 threads, and a meta cache.

This fixes an issue with `CDCService`/xCluster source. The custom client that `CDCService` was using did not have the `SetLocalTabletServer` invoked on it, so it was treating every tablet as a remote tablet. This caused certain optimizations for tablet bootstrapping to get bypassed.

Moved check for `FLAGS_TEST_cdc_inject_replication_index_update_failure` to handle both local and remote tablets.

Fixes #22893
Jira: DB-11795

Test Plan: Jenkins

Reviewers: jhe, stiwary, xCluster

Reviewed By: jhe, stiwary

Subscribers: ybase, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D35906