[#18192] YSQL: Close backend on subtransaction abort failure.
Summary:
D34725 introduced a best effort approach to abort transactions in order to prevent an error stack overflow in case of repeated failures.
This revision extends the same behavior to the abort of subtransactions: if a failure is detected in rolling back to a subtransaction, the backend
connection is terminated. This approach is preferred to handling and propagating the error further because of its simplicity.
This is helpful from an end user's perspective, as the previous approach produced a core-dump (as a result of a PANIC from stack overflow) which
raised a system alert and engaged Support teams for what is an innocuous error. This revision changes the core-dump to a FATAL log message.
Jira: DB-7215

Test Plan:
Manual testing.
Unit tests to follow in a separate revision.

Reviewers: pjain

Reviewed By: pjain

Subscribers: smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35757