[#22948] xCluster: Check replication health at SetupCompletion
Summary:
Make sure replication healthy is OK before succeeding `IsSetupUniverseReplicationDone`.
For the replication healthy to be OK all the Pollers must be in a healthy state.
This allows us to detect and fail the setup in cases such as networking issues between source and target tserver nodes, which is common in Kubernetes setups.

Also fixed an issue in `SyncXClusterConsumerReplicationStatusMap` where a dropped table is never cleaned up from the error map. This was not causing any prod issues since  `PopulateReplicationGroupErrors` skips errors for tables that dont have a stream.

Fixes #22948
Jira: DB-11862

Test Plan: XClusterTest.FailedSetupOnReplicationError

Reviewers: jhe, xCluster, yyan

Reviewed By: yyan

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D35961