[PLAT-14211][PLAT-14300] Simplify master http endpoint + use primary ip for checkUnderReplicatedTablets check
Summary:
Secondary ip is not needed as we are just using the master addresses for http requests and not using it to find the secondary ips from the response. Also simplify the finding of http ports.
Also added check to made sure that CheckUnderreplicatedTablets check is skipped only if the `checkNodesAreSafeToTakeDown` is set to `true`

Test Plan: Create a 6 node rf3 universe. Remove -> Release 2 nodes without a master. Turn off the `checkNodesAreSafeToTakeDown` runtime config. Then perform a stop node on the last node without a master. Make sure that the under-replicated check passes

Reviewers: sanketh

Reviewed By: sanketh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36029